requirements.txt
Django>=6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
openpyxl>=3.1
reportlab>=4.0
docxtpl>=0.16
python-docx>=1.1
docx2pdf>=0.1.8
PyMySQL>=1.1.0
cryptography>=42.0.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
